NoJs version now generates a index.html page
No longer is nojs/pages/* generated. Instead pages are generated to nojs/pageName; where
pageName is the name of the page (with extension changed to html) with the exception that if
the page is named "home.markdown" it will be generated as index.html in the nojs folder.

Fixed links in templates accordingly.