Switch from sass to clay for css pre-processing
Removed dependency on sass in src/site.hs and replaced it with clay.
This is achieved by adding a second executable to the projects cabal
file (called 'gencss') and a folder (called 'clay) that contains all .hs files for
clay (nested however deeply). Then whenever the site is being built, the
"cabal build gencss" command is run, which if successful is followed by
"cabal run gencss compact" whose output is captured and saved to
/default.css.

The sass folder has been left while the majority of the sass has not
been ported over to clay. Only part of sass/default.sass has been ported.

Signed-off-by: Collin J. Doering <collin.doering@rekahsoft.ca>