Modified pagination template
Switched from four .four.column divs (one for firstPage, previousPage,
nextPage, and lastPage) to two .eight.column divs (one for firstPage and
previousPage, and the other for nextPage and lastPage).

This is mostly an ascetic fix that will be modified in the future when I
get around to theming the pagination buttons.

Signed-off-by: Collin J. Doering <collin.doering@rekahsoft.ca>