Fixed opening issue with links within page-content
Previously, all anchor clicks within #page-content was handled by a
event listener that checked whether the anchor's href was to an
'external' url and if so loaded it directly. Otherwise it passed the
value of the anchors href attribute to the jquery address plugin (via
the $.address.value function).

This unfortunately cause clicking on anchors to work fine but when
opening them in a new page or window (or copying the link) caused the
link to reference the html snippet of the page/post instead of the
entire application.

To rectify this, this commit simply transforms internal urls href
attribute after the snippet is loaded into #page-content.

Signed-off-by: Collin J. Doering <collin.doering@rekahsoft.ca>