Upgrade skeleton from 1.2 -> 2.04
Skeleton switched from having 16 columns to 12. A few improvements to
skeleton also required a little tweaking of the current css as well as a
cleanup of templates/default.html.

Note: templates/default-nojs.html was not updated though an update will
be trivial. This is only relevant if the nojs section of the site is
revived.

Signed-off-by: Collin J. Doering <collin.doering@rekahsoft.ca>