Fixed no javascript access message page
When the site is viewed without javascript the navbar forever animates
and the opacity of #page-contents is set, thus making the no javascript
message display not acceptable. To get around this a noscript tag in the
header is added that sets specific CSS in the case the there is indeed
no javascript.

Signed-off-by: Collin J. Doering <collin.doering@rekahsoft.ca>