Remove test suite and update README.md
The test suite depends on hspec-webdriver, which appears to have issues and is not well
supported. Future plans for the site include using hakyll to generate json that will be
consumed by a frontend vuejs application. This will make selenium testing of the hakyll
become absolute (as well as generation of css using clay).

Signed-off-by: Collin J. Doering <collin@rekahsoft.ca>